+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "linter" extension will be documented in this file.
 
+## Unreleased
+
+- Fix swiftlint's ignore file pragma.
+- Fix conditions for ignore code actions.
+
 ## 0.0.0
 
 - Initial release.